Automatically download cef binaries when building
