+++ crates/cef-ui-examples/Cargo.toml
@@ -2,9 +2,13 @@
 name = "cef-ui-examples"
 version = "0.1.0"
 edition = "2021"
+build = "build.rs"
 
 [build-dependencies]
 anyhow = { workspace = true }
+reqwest = { version = "0.12", features = ["blocking"] }
+tar = "0.4"
+flate2 = "1.0"
 
 [dependencies]
 cef-ui = { path = "../cef-ui" }
