+++ crates/cef-ui/Cargo.toml
@@ -13,9 +13,6 @@ tracing-log = "0.2.0"
 parking_lot = "0.12"
 bitflags = "2.4"
 chrono = "0.4"
-reqwest = { version = "0.12", features = ["blocking"] }
-tar = "0.4"
-flate2 = "1.0"
 
 # Bindings for x86_64 Linux.
 [target.'cfg(all(target_os = "linux", target_arch = "x86_64"))'.dependencies]
