+++ crates/cef-ui/src/lib.rs
@@ -1,5 +1,3 @@
 mod cef;
-mod util;
 
 pub use cef::*;
-pub use util::*;
